+++ CHANGELOG.md
@@ -3,13 +3,14 @@
 
 # Unreleased
 
+* Update import-workflow feature (\#269).
 * Revamp home page and user-related layout (\#259).
 * Review logout action (\#259).
 * Implemented editing and deletion of tasks (\#207, \#268). 
-* Bump `word-wrap` from 1.2.3 to 1.2.5 (\#251).
 * Implemented A to B workflow execution (\#254).
-* Improved enable/disable button state (\#257).
 * Fixed workflowtask meta update (\#261).
+* Improved enable/disable button state (\#257).
+* Bump `word-wrap` from 1.2.3 to 1.2.5 (\#251).
 
 # 0.5.2